🩹Item 설정 누락 수정2
