+++ TextRPG_ZOMBIE/TextRPG_ZOMBIE/Item.h
@@ -6,7 +6,7 @@
 
 using namespace std;
 
-enum ItemList // 모든 아이템 목록
+enum ItemType // 모든 아이템 목록
 {
 	Morphine,
 	Epinephrine,
@@ -16,29 +16,40 @@ enum ItemList // 모든 아이템 목록
 	Max
 };
 
-class All_Item // 아이템 설정 및 사용효과
+class Items // 아이템 설정 및 사용효과
 {
 protected:
 	string Name;
-	ItemList ItemType;
+	ItemType itemtype;
 	int Price;
 	int CurrentStack;
 	int MaxStack;
+	// getter setter 필요 상점에서 써야됨
 public:
 
-	All_Item(ItemList type);
+	Items(ItemType type);
 
 	virtual void use(Character* chara);
+
+	// 상점용 getter
+	string getName() const { return Name; }
+	ItemType getItemType() const { return itemtype; }
+	int getPrice() const { return Price; }
+	int getCurrentStack() const { return CurrentStack; }
+	int getMaxStack() const { return MaxStack; }
+
+	// 상점용 setter
+	void setName(const string& newName) { Name = newName; }
+	void setItemType(ItemType newType) { itemtype = newType; }
+	void setPrice(int newPrice) { Price = newPrice; }
+	void setCurrentStack(int newCurrentStack) { CurrentStack = newCurrentStack; }
+	void setMaxStack(int newMaxStack) { MaxStack = newMaxStack; }
 };
 
+// 드랍테이블 컴파일때문에 해놓은 임시 인벤토리
+// 캐릭터 단위에서 인벤토리가 생기면 지울것
 class Inventory
 {
 public:
-	All_Item item[ItemList::Max];
-};
-
-class MonsterDropItem
-{
-public:
-	void dropitem(Inventory* Inven);
+	Items item[ItemType::Max];
 };
