+++ TextRPG_ZOMBIE/TextRPG_ZOMBIE/MonsterDropItem.h
@@ -8,4 +8,3 @@ class MonsterDropItem
 public:
 	void dropItem(Inventory* Inven);
 };
-
